code cleanup
I have:
1.  moved the changes to rollup-specific files
2. marked rollup changes so that it's easier code-search them
