+++ core/types/tx_dynamic_fee.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/math"
-	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/rlp"
 )
 
@@ -104,6 +103,7 @@ func (tx *DynamicFeeTx) effectiveGasPrice(dst *big.Int, baseFee *big.Int) *big.I
 		return dst.Set(tx.GasFeeCap)
 	}
 
+	//[rollup-geth]
 	return dst.Set(math.BigMin(dst.Add(tx.GasTipCap, baseFee), tx.GasFeeCap))
 }
 
@@ -121,24 +121,4 @@ func (tx *DynamicFeeTx) encode(b *bytes.Buffer) error {
 
 func (tx *DynamicFeeTx) decode(input []byte) error {
 	return rlp.DecodeBytes(input, tx)
-}
-
-func (tx *DynamicFeeTx) calldataGas() uint64 {
-	zeroBytes := bytes.Count(tx.Data, []byte{0x00})
-	nonZeroBytes := len(tx.Data) - zeroBytes
-	tokens := uint64(zeroBytes) + uint64(nonZeroBytes)*params.CalldataTokensPerNonZeroByte
-
-	return tokens * params.CalldataGasPerToken
-}
-
-func (tx *DynamicFeeTx) gasLimits() VectorGasLimit {
-	return VectorGasLimit{tx.Gas, 0, tx.calldataGas()}
-}
-
-func (tx *DynamicFeeTx) gasTipCaps() VectorFeeBigint {
-	return VectorFeeBigint{tx.GasTipCap, big.NewInt(0), tx.GasTipCap}
-}
-
-func (tx *DynamicFeeTx) gasFeeCaps() VectorFeeBigint {
-	return VectorFeeBigint{tx.GasFeeCap, big.NewInt(0), tx.GasFeeCap}
 }
